+++ home/runner/work/gradio/gradio/client/js/src/utils/view_api.d.ts
@@ -1,0 +1,2 @@
+import { Client } from "../client";
+export declare function view_api(this: Client): Promise<any>;