+++ home/runner/work/gradio/gradio/js/statustracker/static/Loader.svelte.d.ts
@@ -1,0 +1,6 @@
+interface Props {
+    margin?: boolean;
+}
+declare const Loader: import("svelte").Component<Props, {}, "">;
+type Loader = ReturnType<typeof Loader>;
+export default Loader;