+++ home/runner/work/gradio/gradio/js/statustracker/static/utils.d.ts
@@ -1,0 +1,1 @@
+export declare function pretty_si(num: number): string;