+++ js/component-test/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @self/component-test
 
+## 0.6.0
+
+### Features
+
+- [#13329](https://github.com/gradio-app/gradio/pull/13329) [`7ac583a`](https://github.com/gradio-app/gradio/commit/7ac583a38a4c995d172033ff6e3700390201ff21) - Make builds go zoom zoom.  Thanks @pngwn!
+
+### Dependency updates
+
+- @self/build@0.6.2
+- @gradio/video@0.23.0
+
 ## 0.5.37
 
 ### Dependency updates